Showing the Create and Delete dialogs in front of the content
