+++ packages/www/components/Dashboard/WebhookDialogs/CreateDialog.tsx
@@ -29,21 +29,6 @@ const StyledCrossIcon = styled(Cross1Icon, {
   cursor: "pointer",
 });
 
-const StyledContent = styled(DropdownMenuContent, {
-  minWidth: 400,
-  top: -15,
-  position: "relative",
-  borderLeft: "1px solid $colors$slate7",
-  borderRight: "1px solid $colors$slate7",
-  borderBottom: "1px solid $colors$slate7",
-  backgroundColor: "$loContrast",
-  borderBottomLeftRadius: 6,
-  borderBottomRightRadius: 6,
-  borderTopRightRadius: 0,
-  borderTopLeftRadius: 0,
-  boxShadow: "0 7px 14px 0 rgb(60 66 87 / 8%), 0 0 0 0 rgb(0 0 0 / 12%)",
-});
-
 const eventOptions: Webhook["events"] = [
   "stream.started",
   // "stream.detection", // not yet...
